+++ test/command.js
@@ -4,14 +4,14 @@ var Command = require('../lib/command');
 var commands = require('../lib/commands');
 
 describe('Command', function() {
-  var code, command, sender, target, data, buffer;
+  var code, command, sender, target, data, payload;
 
   beforeEach(function() {
     code = 0x0031;
     data = { level: 100 };
-    buffer = new Buffer([1, 2, 3]);
     sender = { subnet: 1, id: 3 };
     target = { subnet: 1, id: 10 };
+    payload = new Buffer([1, 2, 3]);
 
     simple.mock(sender, 'toString').returnWith('1.3');
     simple.mock(target, 'toString').returnWith('1.10');
@@ -35,19 +35,19 @@ describe('Command', function() {
       });
     });
 
-    it('should set data object', function() {
+    it('should set payload object', function() {
       command = new Command(code, {
-        data: data
+        payload: payload
       });
 
       should(command).have.properties({
-        data: data
+        payload: payload
       });
     });
 
-    it('should skip empty buffer as data', function() {
+    it('should skip empty payload', function() {
       command = new Command(code, {
-        data: new Buffer(0)
+        payload: new Buffer([])
       });
 
       should(command).have.properties({
@@ -69,19 +69,21 @@ describe('Command', function() {
       var command = simple.mock(commands, code, {});
 
       parse = simple.mock(command, 'parse').returnWith(data);
-      encode = simple.mock(command, 'encode').returnWith(buffer);
+      encode = simple.mock(command, 'encode').returnWith(payload);
     });
 
-    it('should parse data buffer', function() {
+    it('should parse data payload', function() {
       command = new Command(code, {
-        data: buffer
+        payload: payload
       });
 
-      should(parse.firstCall.arg).eql(buffer);
+      should(parse.callCount).eql(0);
 
       should(command).have.properties({
         data: data
       });
+
+      should(parse.firstCall.arg).eql(payload);
     });
 
     it('should set parser and encoder', function() {
@@ -101,10 +103,10 @@ describe('Command', function() {
         data: data
       });
 
-      buffer = new Buffer('0E010300000031010A010203', 'hex');
+      var message = new Buffer('0E010300000031010A010203', 'hex');
 
-      should(command.message).eql(buffer);
       should(encode.firstCall.arg).eql(data);
+      should(command.message).eql(message);
     });
 
     it('should encode message without data', function() {
@@ -113,29 +115,45 @@ describe('Command', function() {
         target: target
       });
 
-      buffer = new Buffer('0B010300000031010A', 'hex');
+      var message = new Buffer('0B010300000031010A', 'hex');
 
-      should(command.message).eql(buffer);
+      should(command.message).eql(message);
     });
 
-    it('should have string representaion', function() {
+    it('should have string representaion with payload', function() {
       command = new Command(code, {
         sender: sender,
         target: target,
 
-        data: data
+        payload: payload
       });
 
-      should(command.toString()).eql('1.3 -> 1.10 0x0031: { level: 100 }');
+      should(command.toString()).eql('1.3 -> 1.10 0x0031: 010203 { level: 100 }');
     });
 
-    it('should have string representation without additional data', function() {
+    it('should have string representation without payload', function() {
       command = new Command(0x0004, {
         sender: sender,
         target: target
       });
 
       should(command.toString()).eql('1.3 -> 1.10 0x0004');
+    });
+
+    it('should have string representation with erroneous payload', function() {
+      command = new Command(code, {
+        sender: sender,
+        target: target,
+
+        payload: payload
+      });
+
+      simple.restore(command, 'parse');
+
+      simple.mock(command, 'parse')
+        .throwWith(new Error('Attempt to access memory outside buffer bounds'));
+
+      should(command.toString()).eql('1.3 -> 1.10 0x0031: 010203 PARSE ERROR: Attempt to access memory outside buffer bounds');
     });
   });
 
@@ -144,13 +162,13 @@ describe('Command', function() {
       simple.mock(commands, code, undefined);
     });
 
-    it('should not parse data buffer', function() {
+    it('should not parse payload', function() {
       command = new Command(code, {
-        data: buffer
+        payload: payload
       });
 
       should(command).have.properties({
-        data: buffer
+        data: undefined
       });
     });
 
@@ -159,21 +177,19 @@ describe('Command', function() {
         sender: sender,
         target: target,
 
-        data: buffer
+        payload: payload
       });
 
-      buffer = new Buffer('0E010300000031010A010203', 'hex');
+      var message = new Buffer('0E010300000031010A010203', 'hex');
 
-      should(command.message).eql(buffer);
+      should(command.message).eql(message);
     });
 
     it('should throw error if data is an object', function() {
-      command = new Command(code, {
-        data: data
-      });
-
       should(function() {
-        command.message;
+        command = new Command(code, {
+          data: data
+        });
       }).throw('Data encoder for command 0x0031 is not implemented');
     });
 
@@ -182,7 +198,7 @@ describe('Command', function() {
         sender: sender,
         target: target,
 
-        data: buffer
+        payload: payload
       });
 
       should(command.toString()).eql('1.3 -> 1.10 0x0031: 010203');